+++ components/table/Table.tsx
@@ -2,12 +2,12 @@ import * as React from 'react';
 import classNames from 'classnames';
 import RcTable, { Column, ColumnGroup } from 'rc-table';
 import { TableProps as RcTableProps } from 'rc-table/lib/Table';
-import Checkbox, { CheckboxProps } from '../checkbox';
 import Pagination, { PaginationConfig } from '../pagination';
 import { ConfigContext } from '../config-provider/context';
 import usePagination, { DEFAULT_PAGE_SIZE } from './hooks/usePagination';
 import useLazyKVMap from './hooks/useLazyKVMap';
-import { TableRowSelection, ColumnsType, Key } from './interface';
+import { TableRowSelection, GetRowKey } from './interface';
+import useSelection from './hooks/useSelection';
 
 const EMPTY_LIST: any[] = [];
 
@@ -35,7 +35,7 @@ function Table<RecordType extends object = any>(props: TableProps<RecordType>) {
   const prefixCls = getPrefixCls('table', customizePrefixCls);
 
   // ============================ RowKey ============================
-  const getRowKey = React.useMemo<(record: RecordType, index: number) => Key>(() => {
+  const getRowKey = React.useMemo<GetRowKey<RecordType>>(() => {
     if (typeof rowKey === 'function') {
       return rowKey;
     }
@@ -58,96 +58,12 @@ function Table<RecordType extends object = any>(props: TableProps<RecordType>) {
   }, [mergedData, mergedPagination.current, mergedPagination.pageSize]);
 
   // ========================== Selections ==========================
-  const {
-    selectedRowKeys,
-    getCheckboxProps,
-    onChange: onSelectionChange,
-    columnWidth: selectionColWidth = 60,
-  } = rowSelection || {};
-
-  const [innerSelectedKeys, setInnerSelectedKeys] = React.useState<Key[]>();
-  const mergedSelectedKeys = selectedRowKeys || innerSelectedKeys || EMPTY_LIST;
-
-  const setSelectedKeys = (keys: Key[]) => {
-    setInnerSelectedKeys(keys);
-
-    if (onSelectionChange) {
-      onSelectionChange(keys, keys.map(key => getRecordByKey(key)));
-    }
-  };
-
-  const transformColumns = React.useCallback(
-    (columns: ColumnsType<RecordType>) => {
-      if (!rowSelection) {
-        return columns;
-      }
-
-      // Support selection
-      const keySet = new Set(mergedSelectedKeys);
-
-      // Get all checkbox props
-      const checkboxPropsMap = new Map<Key, Partial<CheckboxProps>>();
-      pageData.forEach((record, index) => {
-        const key = getRowKey(record, index);
-        const checkboxProps = getCheckboxProps ? getCheckboxProps(record) : null;
-        checkboxPropsMap.set(key, checkboxProps || {});
-      });
-
-      // Record key only need check with enabled
-      const recordKeys = pageData
-        .map(getRowKey)
-        .filter(key => !checkboxPropsMap.get(key)!.disabled);
-      const checkedCurrentAll = recordKeys.every(key => keySet.has(key));
-      const checkedCurrentSome = recordKeys.some(key => keySet.has(key));
-
-      const onSelectAllChange = () => {
-        if (checkedCurrentAll) {
-          recordKeys.forEach(key => {
-            keySet.delete(key);
-          });
-        } else {
-          recordKeys.forEach(key => {
-            keySet.add(key);
-          });
-        }
-        setSelectedKeys(Array.from(keySet));
-      };
-
-      return [
-        {
-          width: selectionColWidth,
-          className: `${prefixCls}-selection-column`,
-          title: (
-            <Checkbox
-              checked={checkedCurrentAll}
-              indeterminate={!checkedCurrentAll && checkedCurrentSome}
-              onChange={onSelectAllChange}
-            />
-          ),
-          render: (_: RecordType, record: RecordType, index: number) => {
-            const key = getRowKey(record, index);
-
-            return (
-              <Checkbox
-                {...checkboxPropsMap.get(key)}
-                checked={keySet.has(key)}
-                onChange={() => {
-                  if (keySet.has(key)) {
-                    keySet.delete(key);
-                  } else {
-                    keySet.add(key);
-                  }
-                  setSelectedKeys(Array.from(keySet));
-                }}
-              />
-            );
-          },
-        },
-        ...columns,
-      ];
-    },
-    [getRowKey, pageData, rowSelection, innerSelectedKeys, mergedSelectedKeys, selectionColWidth],
-  );
+  const [transformColumns, selectedKeySet] = useSelection(rowSelection, {
+    prefixCls,
+    data: pageData,
+    getRowKey,
+    getRecordByKey,
+  });
 
   const internalRowClassName = (record: RecordType, index: number, indent: number) => {
     let mergedRowClassName;
@@ -159,7 +75,7 @@ function Table<RecordType extends object = any>(props: TableProps<RecordType>) {
 
     return classNames(
       {
-        [`${prefixCls}-row-selected`]: mergedSelectedKeys.includes(getRowKey(record, index)),
+        [`${prefixCls}-row-selected`]: selectedKeySet.has(getRowKey(record, index)),
       },
       mergedRowClassName,
     );
