+++ components/table/demo/row-selection.md
@@ -8,18 +8,18 @@ only: true
 
 ## zh-CN
 
-第一列是联动的选择框。
+第一列是联动的选择框。可以通过 `rowSelection.type` 属性指定选择类型，默认为 `checkbox`。
 
 > 默认点击 checkbox 触发选择行为，需要点击行触发可以参考例子：<https://codesandbox.io/s/000vqw38rl>
 
 ## en-US
 
-Rows can be selectable by making first column as a selectable column.
+Rows can be selectable by making first column as a selectable column. You can use `rowSelection.type` to set selection type. Default is `checkbox`.
 
-> selection happens when clicking checkbox defaultly. You can see <https://codesandbox.io/s/000vqw38rl> if you need row-click selection behavior.
+> selection happens when clicking checkbox by default. You can see <https://codesandbox.io/s/000vqw38rl> if you need row-click selection behavior.
 
-```jsx
-import { Table } from 'antd';
+```tsx
+import { Table, Radio, Divider } from 'antd';
 
 const columns = [
   {
@@ -74,8 +74,34 @@ const rowSelection = {
   }),
 };
 
-ReactDOM.render(
-  <Table rowSelection={rowSelection} columns={columns} dataSource={data} />,
-  mountNode,
-);
+const Demo = () => {
+  const [selectionType, setSelectionType] = React.useState('checkbox');
+
+  return (
+    <div>
+      <Radio.Group
+        onChange={({ target: { value } }) => {
+          setSelectionType(value);
+        }}
+        value={selectionType}
+      >
+        <Radio value="checkbox">Checkbox</Radio>
+        <Radio value="radio">radio</Radio>
+      </Radio.Group>
+
+      <Divider />
+
+      <Table
+        rowSelection={{
+          type: selectionType,
+          ...rowSelection,
+        }}
+        columns={columns}
+        dataSource={data}
+      />
+    </div>
+  );
+};
+
+ReactDOM.render(<Demo />, mountNode);
 ```
